Count evidence-gate independence in reads, not in labels
An adversarial review of the eval harness found the evidence gate returning
`passed` on an investigation that queried nothing at all. Reproduced against
the real handler and the real gate:

    config + trace     reads=0  refusals=0  gate=passed  reasons=[]
    knowledge + web    reads=0  refusals=0  gate=passed  reasons=[]

Five of the nine signal types have no read ledger behind them, so a check
claiming one is only the model's word. The gate counted them toward "at least
two independent signal types" — the requirement that exists precisely to stop a
verdict resting on the model's own account. Two paragraphs and a straw-man
`ruledOut` produced a verified root cause, and the failure was silent: a passed
gate looks identical whether it was earned or invented.

Filtering by signal type turned out to be the wrong fix. It rejected a real
investigation that ran `ops_run_command` and `metrics_range_query` but tagged
the first check `config` — the work happened; only the label differed. The
ledger's judgment is about the tool, not the label.

So independence is now counted in read families, resolved from the cited tool.
That also closes a second hole nobody had noticed: two calls to `metrics_query`
tagged `metric` and `kubernetes` counted as two independent signals. They are
one family however they are labelled.

    invented tools, unbacked labels  -> unresolved
    same tool, two labels            -> unresolved
    real ops read + real metrics     -> passed

`readFamilyForTool` moves to investigation-state.ts so the gate and the
record-time check cannot drift apart again; when they disagreed, the gate
credited independence the ledger had not verified.

README and the risk model claimed the stronger property this now has. The
prompt tells the model the rule, so it spends turns on a second source rather
than on a second adjective.
