+++ packages/agent-core/src/agent/investigation-state.ts
@@ -9,6 +9,47 @@ export type InvestigationSignalType =
   | 'web'
   | 'other';
 
+/**
+ * The four kinds of read the product can prove actually happened.
+ *
+ * Each maps to a counter in `recordInvestigationRead`, so for a check citing
+ * one of these tools we can say whether a call went out and whether its source
+ * answered. Everything else — a trace viewer, a runbook, a web page — is real
+ * evidence to a human reader but is only the model's word to us.
+ */
+export type ReadFamily = 'metric' | 'log' | 'ops' | 'change';
+
+/**
+ * Which read a check's cited tool draws on, or null when nothing backs it.
+ *
+ * Resolved from the tool name first and the declared signal type only as a
+ * fallback, because the tool is what actually ran and the signal type is a
+ * label the model chose. A check tagged `config` that cites `ops_run_command`
+ * did the work; a check tagged `metric` that cites a tool we have never heard
+ * of did not.
+ *
+ * Shared by the record-time check and the evidence gate on purpose. When these
+ * two disagreed, the gate credited independence the ledger had not verified.
+ */
+export function readFamilyForTool(
+  tool: string,
+  signalType: InvestigationSignalType,
+): ReadFamily | null {
+  const t = tool.trim().toLowerCase();
+  if (t.startsWith('metrics_')) return 'metric';
+  if (t.startsWith('logs_')) return 'log';
+  if (t === 'ops_run_command' || t === 'ops_cluster_shell') return 'ops';
+  if (t === 'changes_list_recent') return 'change';
+  // A tool name the agent invented rather than called. Fall back to the
+  // declared signal type so a check claiming metric evidence still needs a
+  // metric read behind it.
+  if (signalType === 'metric') return 'metric';
+  if (signalType === 'log') return 'log';
+  if (signalType === 'kubernetes') return 'ops';
+  if (signalType === 'change') return 'change';
+  return null;
+}
+
 export type HypothesisStatus = 'supported' | 'ruled_out' | 'inconclusive';
 
 /**
