+++ packages/agent-core/src/agent/orchestrator-prompt.ts
@@ -348,7 +348,7 @@ If the "root cause" is a restated symptom (a status / error code / "the config i
 
 The server enforces a product-agnostic evidence gate on \`investigation_complete\`. A confirmed/likely root-cause claim must satisfy all five checks:
 - direct proof: at least one referenced supported \`investigation_record_check\` directly supports the root-cause object and causal mechanism;
-- cross-checking: \`evidenceRefs\` includes at least two recorded checks across at least two independent signal types;
+- cross-checking: \`evidenceRefs\` includes at least two recorded checks drawing on two *different* read families — \`metrics_*\`, \`logs_*\`, \`ops_run_command\`, \`changes_list_recent\`. The gate resolves the family from the tool the check cites, not from the \`signalType\` label, so two \`metrics_*\` checks are one family however you label them, and a check citing a trace, a config inspection or a runbook counts toward nothing (record it anyway if it is real — it just cannot be what verifies the cause);
 - competing explanations: plausible alternatives were tested and at least one was recorded as \`ruled_out\`;
 - time/scope relevance: at least one referenced check carries \`scope.timeWindow\` or \`scope.affected\` — the gate reads those fields, it does not infer scope from your prose;
 - repair-target consistency: the proposed fix targets the same object/field/mechanism as the proven root cause;
